Create controller file for sending chat to slack
